Go #4: Correct value of firstHalfIndex
The solution was correct before this change, but the value of
firstHalfIndex and the loop condition were incorrect. They were not
working with an index at all, but instead the one-based ordinal of the
middle character.

After this change, firstHalfIndex correctly represents an index and the
loop correctly continues until the character at that index is tested.